refactor: rename indexer-connect to indexer-service
