+++ tools/indexer-service/Dockerfile
