+++ tools/indexer-service/api/buf.yaml
@@ -1,5 +1,5 @@
 version: v1
-name: buf.build/gnolang/gnosocial
+name: buf.build/gnolang/dsocial
 breaking:
   use:
     - FILE
