+++ tools/indexer-service/grpc.go
@@ -10,8 +10,8 @@ import (
 	"connectrpc.com/connect"
 	"connectrpc.com/grpchealth"
 	"connectrpc.com/grpcreflect"
-	api_gen "github.com/gnolang/gnosocial/tools/indexer-connect/api/gen/go"
-	"github.com/gnolang/gnosocial/tools/indexer-connect/api/gen/go/_goconnect"
+	api_gen "github.com/gnolang/gnosocial/tools/indexer-service/api/gen/go"
+	"github.com/gnolang/gnosocial/tools/indexer-service/api/gen/go/_goconnect"
 	"github.com/rs/cors"
 	"go.uber.org/zap"
 	"golang.org/x/net/http2"
