+++ tools/indexer-service/schema/main.go
